Refactor structural hot spots
