+++ apps/desktop/src/main/ipc/register-pipeline-handlers.ts
@@ -4,14 +4,10 @@ import type {
   ActivePipelineSummary,
   HeatmapQueryArgs,
   PipelineRunTimelineEntry,
-  TerminalEventRecord,
-  Thread,
 } from '@shipcode/shared';
 import {
   clampError,
   clarificationAnswerSchema,
-  EXECUTION_PHASES,
-  PAUSABLE_PIPELINE_PHASES,
   PIPELINE_PHASE,
   resolveExecutorModelForIssue,
   resolveRequireApproval,
@@ -19,7 +15,6 @@ import {
   resolveRevisionCount,
   resolveRevisionCountForIssue,
   resolveThreadPhasePresentation,
-  stripAnsi,
 } from '@shipcode/shared';
 
 import { logEvent } from '../logger.service';
@@ -30,180 +25,21 @@ import {
   transitionThreadPhase,
   tryParsePlan,
 } from './helpers';
+import {
+  buildExecutionResumeContext,
+  clampAutoFixFailureOutput,
+  clampResumeText,
+  PAUSABLE_PHASE_SET,
+  RUN_TIMELINE_TERMINAL_LIMIT,
+  STEERING_INSTRUCTION_MAX,
+} from './register-pipeline-handlers.utils';
 import { getRetryAction } from './retry-phase';
 import type { IpcHandlerDeps } from './types';
 
 const execFileAsync = promisify(execFile);
-const AUTO_FIX_FAILURE_OUTPUT_MAX = 8_000;
-const EXECUTION_RESUME_PROMPT_MAX = 24_000;
-const EXECUTION_RESUME_GIT_OUTPUT_MAX = 12_000;
-const EXECUTION_RESUME_TERMINAL_LIMIT = 120;
-const RUN_TIMELINE_TERMINAL_LIMIT = 40;
-const STEERING_INSTRUCTION_MAX = 4_000;
-const INTERRUPTED_EXECUTION_MARKERS = [
-  'interrupted while ShipCode was closed',
-  'interrupted by an app refresh',
-  'Agent process stalled',
-] as const;
-const PAUSABLE_PHASE_SET = new Set<string>(PAUSABLE_PIPELINE_PHASES);
-const EXECUTION_PHASE_SET = new Set<string>(EXECUTION_PHASES);
-
-function clampAutoFixFailureOutput(output: string): string {
-  const cleaned = stripAnsi(output).trim();
-  if (cleaned.length <= AUTO_FIX_FAILURE_OUTPUT_MAX) return cleaned;
-
-  const truncated = cleaned.slice(-AUTO_FIX_FAILURE_OUTPUT_MAX);
-  return `[Earlier terminal output truncated for Auto Fix]\n\n${truncated}`;
-}
-
-function clampResumeText(value: string, max: number, label: string): string {
-  const cleaned = stripAnsi(value).trim();
-  if (cleaned.length <= max) return cleaned;
-  return `${cleaned.slice(0, max)}\n\n[${label} truncated after ${max} chars]`;
-}
-
-function isInterruptedExecutionThread(thread: Thread): boolean {
-  if (
-    thread.status === PIPELINE_PHASE.paused &&
-    thread.pausedPhase &&
-    EXECUTION_PHASE_SET.has(thread.pausedPhase)
-  ) {
-    return true;
-  }
-  if (
-    thread.failurePhase !== PIPELINE_PHASE.executing &&
-    thread.status !== PIPELINE_PHASE.executing
-  ) {
-    return false;
-  }
-  return INTERRUPTED_EXECUTION_MARKERS.some((marker) => thread.lastError?.includes(marker));
-}
-
 function throwClampedIpcError(scope: string, error: unknown): never {
   console.error(`[ipc] ${scope} failed`, error);
   throw new Error(clampError(error));
-}
-
-function formatTerminalResumeLine(record: TerminalEventRecord): string | null {
-  const event = record.event;
-  switch (event.kind) {
-    case 'user_input':
-      return `[user] ${event.content}`;
-    case 'text':
-    case 'raw':
-    case 'thinking':
-      return event.content;
-    case 'error':
-      return `[error] ${event.message}`;
-    case 'lifecycle':
-      return `[lifecycle] ${event.message}`;
-    case 'tool_start':
-      return `[tool:start] ${event.name}: ${event.summary}`;
-    case 'tool_end':
-      return `[tool:end] ${event.name}${event.exitCode === undefined ? '' : ` exit=${event.exitCode}`}${
-        event.outputSummary ? ` ${event.outputSummary}` : ''
-      }`;
-    case 'turn_start':
-      return `[turn:start] ${event.turn}`;
-    case 'turn_end':
-      return `[turn:end] ${event.turn}`;
-    case 'action':
-      return `[action] ${event.label}`;
-    case 'clarification_requested':
-      return `[clarification_requested] ${event.summary}`;
-    case 'clarification_answered':
-      return `[clarification_answered] ${event.questionCount} answers`;
-    case 'done':
-      return '[done]';
-    default:
-      return null;
-  }
-}
-
-async function readGitResumeOutput(cwd: string, args: string[], label: string): Promise<string> {
-  try {
-    const { stdout } = await execFileAsync('git', args, {
-      cwd,
-      encoding: 'utf8',
-      maxBuffer: EXECUTION_RESUME_GIT_OUTPUT_MAX * 4,
-    });
-    return clampResumeText(String(stdout), EXECUTION_RESUME_GIT_OUTPUT_MAX, label);
-  } catch (error) {
-    const message = error instanceof Error ? error.message : String(error);
-    return `[Unable to read ${label}: ${message}]`;
-  }
-}
-
-async function buildExecutionResumeContext(
-  thread: Thread,
-  queries: IpcHandlerDeps['queries'],
-): Promise<string | null> {
-  if (!isInterruptedExecutionThread(thread) || !thread.worktreePath) return null;
-
-  const isPaused = thread.status === PIPELINE_PHASE.paused;
-  const interruptedPhase = isPaused ? thread.pausedPhase : (thread.failurePhase ?? thread.status);
-  const checkpoint = queries.checkpoints.getLatest(thread.id);
-  const diffBase = checkpoint?.commitSha ?? 'HEAD';
-  const [status, changedFiles, diffStat, diffExcerpt] = await Promise.all([
-    readGitResumeOutput(thread.worktreePath, ['status', '--short'], 'git status'),
-    readGitResumeOutput(
-      thread.worktreePath,
-      ['diff', '--name-only', diffBase, '--'],
-      'changed files',
-    ),
-    readGitResumeOutput(thread.worktreePath, ['diff', '--stat', diffBase, '--'], 'diff stat'),
-    readGitResumeOutput(thread.worktreePath, ['diff', diffBase, '--'], 'diff excerpt'),
-  ]);
-  const terminalTail = queries.terminalEvents
-    .listByThread(thread.id, EXECUTION_RESUME_TERMINAL_LIMIT)
-    .flatMap((event) => {
-      const line = formatTerminalResumeLine(event);
-      return line?.trim() ? [line] : [];
-    })
-    .join('\n');
-
-  const sections = [
-    '<interrupted_run_context>',
-    isPaused
-      ? 'The previous execution was paused by the user. This is a semantic resume, not a clean restart.'
-      : 'The previous execution was interrupted. This is a semantic resume, not a clean restart.',
-    '',
-    `Interrupted phase: ${interruptedPhase ?? thread.status}`,
-    `Last error: ${thread.lastError ?? (isPaused ? 'Paused by user' : 'Unknown interruption')}`,
-    `Worktree: ${thread.worktreePath}`,
-    checkpoint
-      ? `Checkpoint before execution: ${checkpoint.label} (${checkpoint.commitSha.slice(0, 12)})`
-      : 'Checkpoint before execution: none recorded; using HEAD as the diff base.',
-    '',
-    '<current_git_status>',
-    status || 'Clean',
-    '</current_git_status>',
-    '',
-    '<changed_files_since_checkpoint>',
-    changedFiles || 'No changed files detected',
-    '</changed_files_since_checkpoint>',
-    '',
-    '<diff_stat_since_checkpoint>',
-    diffStat || 'No diff stat available',
-    '</diff_stat_since_checkpoint>',
-    '',
-    '<diff_excerpt_since_checkpoint>',
-    diffExcerpt || 'No diff available',
-    '</diff_excerpt_since_checkpoint>',
-    '',
-    '<terminal_tail>',
-    clampResumeText(terminalTail || 'No terminal events recorded', 8_000, 'terminal tail'),
-    '</terminal_tail>',
-    '',
-    'Resume instructions:',
-    '- Treat the current worktree as WIP from the interrupted agent.',
-    '- Inspect and preserve useful changes before continuing.',
-    '- Continue the approved plan from remaining work; fix incomplete or broken edits.',
-    '- Do not reset or recreate the worktree unless the WIP is clearly wrong.',
-    '</interrupted_run_context>',
-  ].join('\n');
-
-  return `\n\n${clampResumeText(sections, EXECUTION_RESUME_PROMPT_MAX, 'resume context')}`;
 }
 
 export function registerPipelineHandlers({
@@ -708,7 +544,7 @@ export function registerPipelineHandlers({
         (active.activeProcessId ? processManager.get(active.activeProcessId) : undefined) ??
         processManager.listActive().find((process) => process.threadId === threadId);
 
-      if (!liveProcess || liveProcess.state !== 'running') {
+      if (liveProcess?.state !== 'running') {
         return {
           threadId,
           status: 'stale',
