+++ apps/desktop/src/renderer/components/IssueDetail.tsx
@@ -66,7 +66,7 @@ import {
   Play,
   Square,
 } from 'lucide-react';
-import { useCallback, useEffect, useMemo, useReducer, useRef, useState } from 'react';
+import { useEffect, useMemo, useReducer, useRef, useState } from 'react';
 import { STABLE_APP_STATE_STALE_TIME } from '../query-stale-times';
 import { useAppStore } from '../stores/app-store';
 import { toast } from '../stores/toast-store';
@@ -76,16 +76,19 @@ import {
   decodePhaseOption,
   encodePhaseOption,
   resolveFailingPhaseOutput,
+  resolveIssueRetryPresentation,
 } from './issue-detail/helpers';
 import { buildIssueDetailActions } from './issue-detail/IssueDetailActions';
 import { buildIssueDetailDialogs } from './issue-detail/IssueDetailDialogs';
 import { IssueDetailTabs } from './issue-detail/IssueDetailTabs';
 import { PipelineTab } from './issue-detail/PipelineTab';
 import type { IssueDetailTab } from './issue-detail/tab-types';
+import {
+  DETAIL_SIDEBAR_MAX,
+  DETAIL_SIDEBAR_MIN,
+  useResizableDetailSidebar,
+} from './issue-detail/useResizableDetailSidebar';
 
-const DETAIL_SIDEBAR_MIN = 320;
-const DETAIL_SIDEBAR_MAX = 640;
-const DETAIL_SIDEBAR_DEFAULT = 416; // matches previous w-[26rem]
 const EMPTY_PLAN_HISTORY: PlanRecord[] = [];
 
 const INHERIT_EXECUTOR_VALUE = '__inherit__';
@@ -179,36 +182,7 @@ function useIssueDetailView() {
     };
   }, []);
 
-  // Detail sidebar resize (matches TerminalDrawer / ProjectSidebar pattern)
-  const [detailSidebarWidth, setDetailSidebarWidth] = useState(DETAIL_SIDEBAR_DEFAULT);
-  const detailDragRef = useRef<{ startX: number; startW: number } | null>(null);
-  const handleDetailResizeMouseDown = useCallback(
-    (e: React.MouseEvent) => {
-      e.preventDefault();
-      detailDragRef.current = { startX: e.clientX, startW: detailSidebarWidth };
-      const onMove = (ev: MouseEvent) => {
-        const drag = detailDragRef.current;
-        if (!drag) return;
-        // Dragging left = wider sidebar (inverted delta)
-        const delta = drag.startX - ev.clientX;
-        const next = Math.min(
-          DETAIL_SIDEBAR_MAX,
-          Math.max(DETAIL_SIDEBAR_MIN, drag.startW + delta),
-        );
-        setDetailSidebarWidth(next);
-      };
-      const onUp = () => {
-        detailDragRef.current = null;
-        document.removeEventListener('mousemove', onMove);
-        document.removeEventListener('mouseup', onUp);
-        document.body.classList.remove('cursor-col-resize', 'select-none');
-      };
-      document.addEventListener('mousemove', onMove);
-      document.addEventListener('mouseup', onUp);
-      document.body.classList.add('cursor-col-resize', 'select-none');
-    },
-    [detailSidebarWidth],
-  );
+  const { detailSidebarWidth, handleDetailResizeMouseDown } = useResizableDetailSidebar();
 
   const [phaseModelValidation, setPhaseModelValidation] = useState<
     Partial<
@@ -535,51 +509,10 @@ function useIssueDetailView() {
     () => normalizedThreadPlanHistory[0] ?? null,
     [normalizedThreadPlanHistory],
   );
-  const failedLatestStructuredVerification =
-    latestVerification?.planId === latestPlan?.id &&
-    latestVerification?.result === 'failed' &&
-    !!latestVerification?.structured;
-  const retryAction = useMemo(() => {
-    if (!thread) return null;
-    if (/no code changes/i.test(thread.lastError ?? '')) return 'plan' as const;
-    const structuredPlan = latestPlan?.structured ?? null;
-    if (!structuredPlan) return 'plan' as const;
-    if (!thread.worktreePath) return 'review' as const;
-    if (latestVerification && latestVerification.planId === latestPlan?.id) {
-      if (latestVerification.result === 'failed' && latestVerification.structured) {
-        return 'execute' as const;
-      }
-      if (latestVerification.result === 'failed') return 'verify' as const;
-      if (latestVerification.result === 'passed') return 'commit_and_push' as const;
-    }
-    return 'execute' as const;
-  }, [latestPlan?.id, latestPlan?.structured, latestVerification, thread]);
-  const retryButtonLabel =
-    retryAction === 'review'
-      ? 'Resume review'
-      : retryAction === 'execute'
-        ? 'Resume execution'
-        : retryAction === 'verify'
-          ? 'Resume verification'
-          : retryAction === 'commit_and_push'
-            ? 'Resume shipping'
-            : 'Re-plan';
-  const retrySummary =
-    retryAction === 'review'
-      ? 'Retry will resume from review using the latest structured plan.'
-      : retryAction === 'execute'
-        ? failedLatestStructuredVerification
-          ? 'Retry will resume from execution using the current worktree and latest verification feedback.'
-          : 'Retry will resume from execution using the latest structured plan.'
-        : retryAction === 'verify'
-          ? 'Retry will resume from verification using the current worktree.'
-          : retryAction === 'commit_and_push'
-            ? 'Retry will resume from commit and push using the verified worktree.'
-            : retryAction === 'plan'
-              ? /no code changes/i.test(thread?.lastError ?? '')
-                ? 'The executor produced no file changes. Update the issue description with more detail before replanning.'
-                : 'Retry will start a fresh planning pass. This resumes the workflow, not the same live planner session.'
-              : null;
+  const { retryButtonLabel, retrySummary } = useMemo(
+    () => resolveIssueRetryPresentation({ thread, latestPlan, latestVerification }),
+    [latestPlan, latestVerification, thread],
+  );
   const threadPhase = currentPipelinePhase;
   const canStartPipeline =
     !activeThreadId &&
