+++ packages/shared/src/types/project.ts
@@ -1,0 +1,104 @@
+import type { AgentType, PipelineSpeedProfile, ReasoningEffort, RevisionCount } from './agents';
+import type { GhStatusMapping } from './github';
+import type { ProjectNotificationRoutingMode, ProjectSetupStatus } from './settings';
+
+// === Project Types ===
+
+export interface Project {
+  id: string;
+  name: string;
+  path: string;
+  pathExists?: boolean;
+  setupStatus?: ProjectSetupStatus;
+  setupPath?: string;
+  setupError?: string | null;
+  gitRemote: string | null;
+  githubRepoId: string | null;
+  githubRepoFullName: string | null;
+  starterIssueNumber: number | null;
+  starterIssueCreatedAt: string | null;
+  /**
+   * Optional override for the Kanban `board` quick-link. GitHub Projects v2
+   * live under a user/org (`/users/<name>/projects/<n>` or `/orgs/<name>/projects/<n>`)
+   * and can span multiple repos, so we can't derive this from `gitRemote` alone.
+   * When null, the Kanban header hides the `board` quick-link.
+   */
+  githubProjectUrl: string | null;
+  /**
+   * Per-project mapping from ShipCode macro columns to GH Projects v2 Status
+   * option names. Auto-detected when `githubProjectUrl` is set. When null,
+   * GH Status sync is disabled for this project.
+   */
+  githubStatusMapping: GhStatusMapping | null;
+  plannerModelOverride: AgentType | null;
+  reviewerModelOverride: AgentType | null;
+  executorModelOverride: AgentType | null;
+  verifierModelOverride: AgentType | null;
+  plannerModelIdOverride: string | null;
+  reviewerModelIdOverride: string | null;
+  executorModelIdOverride: string | null;
+  verifierModelIdOverride: string | null;
+  plannerReasoningEffortOverride: ReasoningEffort | null;
+  reviewerReasoningEffortOverride: ReasoningEffort | null;
+  executorReasoningEffortOverride: ReasoningEffort | null;
+  verifierReasoningEffortOverride: ReasoningEffort | null;
+  revisionCountOverride: RevisionCount | null;
+  requireApprovalOverride?: boolean | null;
+  pipelineSpeedProfileOverride?: PipelineSpeedProfile | null;
+  prdQualityGate?: boolean | null;
+  discordRouting: ProjectNotificationRoutingMode;
+  discordWebhookUrlOverride: string | null;
+  telegramRouting: ProjectNotificationRoutingMode;
+  telegramChatIdOverride: string | null;
+  defaultBranch: string;
+  pinned: boolean;
+  archived: boolean;
+  hidden: boolean;
+  notifyGithubUser: string | null;
+  createdAt: string;
+  updatedAt: string;
+}
+
+// === Automation Types ===
+
+export type AutomationLastStatus = 'running' | 'completed' | 'failed';
+
+export interface Automation {
+  id: string;
+  projectId: string;
+  name: string;
+  prompt: string;
+  cronExpr: string;
+  enabled: boolean;
+  executorProvider: AgentType | null;
+  executorModelId: string | null;
+  executorReasoningEffort: ReasoningEffort | null;
+  lastStartedAt: string | null;
+  lastCompletedAt: string | null;
+  lastStatus: AutomationLastStatus | null;
+  nextRunAt: string | null;
+  runCount: number;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface CreateAutomationInput {
+  projectId: string;
+  name: string;
+  prompt: string;
+  cronExpr: string;
+  enabled?: boolean;
+  executorProvider?: AgentType | null;
+  executorModelId?: string | null;
+  executorReasoningEffort?: ReasoningEffort | null;
+}
+
+export interface UpdateAutomationInput {
+  name?: string;
+  prompt?: string;
+  cronExpr?: string;
+  enabled?: boolean;
+  executorProvider?: AgentType | null;
+  executorModelId?: string | null;
+  executorReasoningEffort?: ReasoningEffort | null;
+}
